Move DoubleArrayAhoCorasick struct into bytewise module
