+++ src/charwise/builder.rs
@@ -34,7 +34,7 @@ impl CharwiseDoubleArrayAhoCorasickBuilder {
     /// # Examples
     ///
     /// ```
-    /// use daachorse::charwise::CharwiseDoubleArrayAhoCorasickBuilder;
+    /// use daachorse::CharwiseDoubleArrayAhoCorasickBuilder;
     ///
     /// let patterns = vec!["全世界", "世界", "に"];
     ///
@@ -111,7 +111,7 @@ impl CharwiseDoubleArrayAhoCorasickBuilder {
     /// # Examples
     ///
     /// ```
-    /// use daachorse::charwise::CharwiseDoubleArrayAhoCorasickBuilder;
+    /// use daachorse::CharwiseDoubleArrayAhoCorasickBuilder;
     ///
     /// let patterns = vec!["全世界", "世界", "に"];
     /// let pma = CharwiseDoubleArrayAhoCorasickBuilder::new().build(patterns).unwrap();
@@ -159,7 +159,7 @@ impl CharwiseDoubleArrayAhoCorasickBuilder {
     /// # Examples
     ///
     /// ```
-    /// use daachorse::charwise::CharwiseDoubleArrayAhoCorasickBuilder;
+    /// use daachorse::CharwiseDoubleArrayAhoCorasickBuilder;
     ///
     /// let patvals = vec![("全世界", 0), ("世界", 10), ("に", 100)];
     /// let pma = CharwiseDoubleArrayAhoCorasickBuilder::new().build_with_values(patvals).unwrap();
