+++ src/charwise/iter.rs
@@ -4,8 +4,9 @@ use core::iter::Enumerate;
 use core::num::NonZeroU32;
 
 use crate::charwise::CharwiseDoubleArrayAhoCorasick;
+
+use crate::charwise::ROOT_STATE_IDX;
 use crate::Match;
-use crate::ROOT_STATE_IDX;
 
 /// Iterator for some struct that implements [`AsRef<str>`].
 pub struct StrIterator<P> {
